+++ R/dates.R
@@ -40,16 +40,3 @@ setMethod(
     c(lower = inf, upper = sup, duration = sup - inf)
   }
 )
-
-# Older ========================================================================
-#' @export
-#' @rdname older
-#' @aliases test_older,MCMC-method
-setMethod(
-  f = "test_older",
-  signature = c(x = "numeric", y = "numeric"),
-  definition = function(x, y, ...) {
-    ## Bayesian test : x < y
-    mean(x < y)
-  }
-)