+++ inst/examples/ex-coerce.R
@@ -1,5 +1,2 @@
 ## Coerce to MCMC
 eve <- as_mcmc(events, iteration = 1)
-
-## Coerce to Phases
-pha <- as_phases(phases, start = c(1, 3), iteration = 1)