+++ inst/examples/ex-range.R
@@ -4,7 +4,7 @@ eve <- eve[1:10000, ]
 
 ## Compute min-max range by group
 ## Unless otherwise specified, the phases are assumed to be unordered
-pha <- phase(eve, groups = list(A = c(1, 3), B = c(2, 4)))
+pha <- as_phases(eve, groups = list(A = c(1, 3), B = c(2, 4)))
 
 ## Set chronological order
 ## (from the oldest to the youngest phase)